// src/models/liturgy-masters/LiturgyMasters.ts
export default class LiturgyMasters {
  OfficeCommonPsalter: OfficeCommonPsalter;
  InvitationCommonPsalter: InvitationCommonPsalter;
  OfficeOfOrdinaryTime: OfficeOfOrdinaryTime;
  PrayersOfOrdinaryTime: PrayersOfOrdinaryTime;
  PrayersOfOrdinaryTimeWhenFirstVespers: PrayersOfOrdinaryTime;
  CommonPartsUntilFifthWeekOfLentTime: CommonPartsUntilFifthWeekOfLentTime;
  PartsOfLentTime: PartsOfLentTime;
  PartsOfFiveWeeksOfLentTime: PartsOfFiveWeeksOfLentTime;
  CommonPartsOfHolyWeek: CommonPartsOfHolyWeek;
  PalmSundayParts: PalmSundayParts;
  PartsOfHolyWeek: PartsOfHolyWeek;
  PartsOfEasterTriduum: PartsOfEasterTriduum;
  PartsOfEasterBeforeAscension: PartsOfEasterBeforeAscension;
  PartsOfEasterOctave: PartsOfEasterOctave;
  PartsOfEasterAfterAscension: PartsOfEasterAfterAscension;
  EasterWeekParts: EasterWeekParts;
  CommonAdventAndChristmasParts: CommonAdventAndChristmasParts;
  AdventWeekParts: AdventWeekParts;
  AdventSundayParts: AdventSundayParts;
  AdventFirstVespersOfSundayParts: AdventSundayParts;
  AdventFairDaysParts: AdventFairDaysParts;
  AdventFairDaysAntiphons: AdventFairDaysAntiphons;
  ChristmasWhenOctaveParts: ChristmasWhenOctaveParts;
  ChristmasBeforeEpiphanyParts: ChristmasBeforeEpiphanyParts;
  SpecialCommonPartsOfEasterSundays: SpecialCommonPartsOfEasterSundays;
  LaudesCommonPsalter: LaudesCommonPsalter;
  CommonSpecialPartsOfEaster: CommonSpecialPartsOfEaster;
  EasterSundayParts: EasterSundayParts;
  EasterFirstVespersOfSundayParts: EasterSundayParts;
  EasterSunday: EasterSunday;
  FiveWeeksOfSundayLentParts: FiveWeeksOfSundayLentParts;
  FiveWeeksOfFirstsVespersOfSundayLentParts: FiveWeeksOfSundayLentParts;
  VespersCommonPsalter: VespersCommonPsalter;
  SolemnityAndFestivityParts: SolemnityAndFestivityParts;
  SolemnityAndFestivityWhenFirstVespersParts: SolemnityAndFestivityParts;
  CommonHourPsalter: CommonHourPsalter;
  CommonNightPrayerPsalter: CommonNightPrayerPsalter;
  CommonOfficeWhenStrongTimesPsalter: CommonOfficeWhenStrongTimesPsalter;
  SaintsSolemnities: SaintsSolemnities;
  SaintsSolemnitiesWhenFirstsVespersParts: SaintsSolemnities;
  SaintsMemories: SaintsMemories;
  SpecialDaysParts: SpecialDaysParts;
  Various: Various;
}